bidirectional: combine both sign respond and read respond
