+++ chain-signatures/node/src/lib.rs
@@ -10,6 +10,7 @@ pub mod mesh;
 pub mod metrics;
 pub mod node_client;
 pub mod protocol;
+pub mod read_respond;
 pub mod rpc;
 pub mod sign_respond_tx;
 pub mod storage;
