+++ chain-signatures/node/src/rpc.rs
@@ -1310,13 +1310,13 @@ async fn try_publish_sol(
         })?;
 
     tracing::info!(
-        sign_id = ?action.request.indexed.id,
-        tx_hash = ?tx,
-        elapsed = ?timestamp.elapsed(),
-        "published solana signature successfully"
+    sign_id = ?action.request.indexed.id,
+    tx_hash = ?tx,
+    elapsed = ?timestamp.elapsed(),
+    "published solana signature successfully"
     );
 
-    if let SignRequestType::SignRespond = action.request.indexed.sign_request_type {
+    if let SignRequestType::SignRespond(_) = action.request.indexed.sign_request_type {
         if sign_respond_responded_send
             .send((action.clone(), original_signature))
             .await
